branch: rename orphan branches in any worktree
In cfaff3aac (branch -m: allow renaming a yet-unborn branch, 2020-12-13)
we added support for renaming an orphan branch in the current working
tree, avoiding the rename_ref() error because an orphan branch does not
point to any ref.

We did this with two checks (note that head cannot be null due to a
precondition in cmd_branch):

 	 if (branch_is_not_head)
 	 	// Only HEAD can be orphan. Do the rename_ref
 	 else if (head_ref_is_valid)
 		// There is a valid ref. Do the rename_ref
 	 else
 		// Nothing to do here. Continue

With worktrees, we can try to rename an orphan branch while we are in a
different worktree, and so the orphan branch we want to rename is not
the HEAD in the current worktree:

	$ git worktree add orphan-worktree --detach
	$ git -C orphan-worktree checkout --orphan orphn-branch
	$ git branch -m orphn-branch orphan-branch
	fatal: No branch named 'orphn-branch'.

Lets fix this considering all HEADs in all worktrees in the repository,
changing the checks introduced in cfaff3aac8 to:

 	 if (branch_is_not_head_in_any_worktree)
 	 	// Only HEADs can be orphan. Do the rename_ref
 	 else if (branch_ref_is_valid)
 		// There is a valid ref. Do the rename_ref
 	 else
 		// Nothing to do here. Continue

Let's also add to t3200-branch tests for this and for normal renames of
orphan branches (other than the initial branch) for which we did not yet
have tests.

Signed-off-by: Rubén Justo <rjusto@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>